refactor(steplibrary): split steplibrary.go by concern
steplib.go  — Steplib struct, New, v2CacheDir
activate.go — Activate
resolve.go  — getStepVersionInfo, resolveMinorLocked, toStepGroupInfoModel

toStepGroupInfoModel lives with its only caller (getStepVersionInfo) in
resolve.go rather than activate.go.
